requirements.txt
spotipy
python-dotenv
wordcloud
matplotlib